Fix Razorpay amount casting and add detailed error response

+++ foodiesapi/src/main/java/in/bushansirgur/foodiesapi/service/OrderServiceImpl.java
@@ -43,7 +43,7 @@ public class OrderServiceImpl implements OrderService{
         try {
             RazorpayClient razorpayClient = new RazorpayClient(RAZORPAY_KEY, RAZORPAY_SECRET);
             JSONObject orderRequest = new JSONObject();
-            orderRequest.put("amount", newOrder.getAmount() * 100);
+            orderRequest.put("amount", (int) (newOrder.getAmount() * 100));
             orderRequest.put("currency", "INR");
             orderRequest.put("payment_capture", 1);
